[ci] Move one test to the flaky suite
