+++ pulsar-broker/src/test/java/org/apache/pulsar/broker/service/plugin/FilterEntryTest.java
@@ -34,7 +34,6 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
 
 import lombok.extern.slf4j.Slf4j;
@@ -50,7 +49,6 @@ import org.apache.pulsar.client.api.Consumer;
 import org.apache.pulsar.client.api.Message;
 import org.apache.pulsar.client.api.Producer;
 import org.apache.pulsar.client.api.Schema;
-import org.apache.pulsar.client.api.SubscriptionType;
 import org.apache.pulsar.client.impl.MessageIdImpl;
 import org.apache.pulsar.common.nar.NarClassLoader;
 import org.apache.pulsar.common.stats.AnalyzeSubscriptionBacklogResult;
@@ -322,119 +320,6 @@ public class FilterEntryTest extends BrokerTestBase {
             assertEquals(filtered, 10);
         }
     }
-
-    @Test
-    public void testEntryFilterRescheduleMessageDependingOnConsumerSharedSubscription() throws Throwable {
-        String topic = "persistent://prop/ns-abc/topic" + UUID.randomUUID();
-        String subName = "sub";
-
-        Map<String, String> metadataConsumer1 = new HashMap<>();
-        metadataConsumer1.put("matchValueAccept", "FOR-1");
-        metadataConsumer1.put("matchValueReschedule", "FOR-2");
-
-        Map<String, String> metadataConsumer2 = new HashMap<>();
-        metadataConsumer2.put("matchValueAccept", "FOR-2");
-        metadataConsumer2.put("matchValueReschedule", "FOR-1");
-        final int numMessages = 200;
-
-        try (Producer<String> producer = pulsarClient.newProducer(Schema.STRING)
-                .enableBatching(false).topic(topic).create();
-             Consumer<String> consumer1 = pulsarClient.newConsumer(Schema.STRING).topic(topic)
-                     .subscriptionType(SubscriptionType.Shared)
-                     .properties(metadataConsumer1)
-                     .consumerName("consumer1")
-                     .receiverQueueSize(numMessages / 2)
-                     .subscriptionName(subName)
-                     .subscribe();
-             Consumer<String> consumer2 = pulsarClient.newConsumer(Schema.STRING)
-                     .subscriptionType(SubscriptionType.Shared)
-                     .properties(metadataConsumer2)
-                     .consumerName("consumer2")
-                     .topic(topic)
-                     .receiverQueueSize(numMessages / 2)
-                     .subscriptionName(subName)
-                     .subscribe()) {
-
-            // mock entry filters
-            PersistentSubscription subscription = (PersistentSubscription) pulsar.getBrokerService()
-                    .getTopicReference(topic).get().getSubscription(subName);
-            Dispatcher dispatcher = subscription.getDispatcher();
-            Field field = EntryFilterSupport.class.getDeclaredField("entryFilters");
-            field.setAccessible(true);
-            NarClassLoader narClassLoader = mock(NarClassLoader.class);
-            EntryFilter filter1 = new EntryFilterTest();
-            EntryFilterWithClassLoader loader1 =
-                    spyWithClassAndConstructorArgs(EntryFilterWithClassLoader.class, filter1, narClassLoader);
-            EntryFilter filter2 = new EntryFilterTest();
-            EntryFilterWithClassLoader loader2 =
-                    spyWithClassAndConstructorArgs(EntryFilterWithClassLoader.class, filter2, narClassLoader);
-            field.set(dispatcher, ImmutableList.of(loader1, loader2));
-
-            for (int i = 0; i < numMessages; i++) {
-                if (i % 2 == 0) {
-                    producer.newMessage()
-                            .property("FOR-1", "")
-                            .value("consumer-1")
-                            .send();
-                } else {
-                    producer.newMessage()
-                            .property("FOR-2", "")
-                            .value("consumer-2")
-                            .send();
-                }
-            }
-            CompletableFuture<Void> resultConsume1 = new CompletableFuture<>();
-            pulsar.getExecutor().submit(() -> {
-                try {
-                    // assert that the consumer1 receive all the messages and that such messages
-                    // are for consumer1
-                    int counter = 0;
-                    while (counter < numMessages / 2) {
-                        Message<String> message = consumer1.receive(1, TimeUnit.MINUTES);
-                        if (message != null) {
-                            log.info("received1 {} - {}", message.getValue(), message.getProperties());
-                            counter++;
-                            assertEquals("consumer-1", message.getValue());
-                            consumer1.acknowledgeAsync(message);
-                        } else {
-                            resultConsume1.completeExceptionally(
-                                    new Exception("consumer1 did not receive all the messages"));
-                        }
-                    }
-                    resultConsume1.complete(null);
-                } catch (Throwable err) {
-                    resultConsume1.completeExceptionally(err);
-                }
-            });
-
-            CompletableFuture<Void> resultConsume2 = new CompletableFuture<>();
-            pulsar.getExecutor().submit(() -> {
-                try {
-                    // assert that the consumer2 receive all the messages and that such messages
-                    // are for consumer2
-                    int counter = 0;
-                    while (counter < numMessages / 2) {
-                        Message<String> message = consumer2.receive(1, TimeUnit.MINUTES);
-                        if (message != null) {
-                            log.info("received2 {} - {}", message.getValue(), message.getProperties());
-                            counter++;
-                            assertEquals("consumer-2", message.getValue());
-                            consumer2.acknowledgeAsync(message);
-                        } else {
-                            resultConsume2.completeExceptionally(
-                                    new Exception("consumer2 did not receive all the messages"));
-                        }
-                    }
-                    resultConsume2.complete(null);
-                } catch (Throwable err) {
-                    resultConsume2.completeExceptionally(err);
-                }
-            });
-            resultConsume1.get(1, TimeUnit.MINUTES);
-            resultConsume2.get(1, TimeUnit.MINUTES);
-        }
-    }
-
 
     private void verifyBacklog(String topic, String subscription,
                                int numEntries, int numMessages,
